requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databases==0.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.4.8
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
